Add migration

+++ app/commands/remove.py
@@ -8,6 +8,7 @@ from ..utils import format_currency, parse_int
 
 def remove(update: Update, context: CallbackContext):
   chat_id = update.message.chat_id
+  user_id = update.message.from_user.id
 
   try:
     expense_id = parse_int(context.args[0])
@@ -22,17 +23,22 @@ def remove(update: Update, context: CallbackContext):
       quote=True,
     )
 
-  reply = f'👌 Removed expense for @{expense.user} {format_currency(expense.amount)}'
+  reply = ''
 
-  if expense.label.strip():
-    reply += f': {expense.label}'
+  if expense.user_id == user_id:
+    reply = f'👌 Removed expense for {expense.user_alias} {format_currency(expense.amount)}'
 
-  try:
-    db.session.delete(expense)
-    db.session.commit()
-  except Exception as exc:
-    db.session.rollback()
-    raise exc
+    if expense.label.strip():
+      reply += f': {expense.label}'
+
+    try:
+      db.session.delete(expense)
+      db.session.commit()
+    except Exception as exc:
+      db.session.rollback()
+      raise exc
+  else:
+    reply = f'✋ You can\'t remove someone else\'s expense, have {expense.user_alias} do it instead'
 
   update.message.reply_markdown(
     reply,
